+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/expression/ConditionalExpression.java
@@ -1,6 +1,5 @@
 package com.dat3m.dartagnan.program.expression;
 
-import com.dat3m.dartagnan.expression.processing.ExpressionVisitor;
 import com.dat3m.dartagnan.program.Register;
 import com.google.common.collect.ImmutableSet;
